refact!

+++ ACL/src/ACL.Business/Application/Interfaces/Repositories/IBranchRepository.cs
@@ -8,7 +8,7 @@ namespace ACL.Business.Application.Interfaces.Repositories
         /// <inheritdoc/>
         IEnumerable<Branch>? All();
         /// <inheritdoc/>
-        Branch? GetById(ulong id);
+        Branch? GetById(uint id);
         /// <inheritdoc/>
         Branch? Add(Branch entity);
         /// <inheritdoc/>
@@ -16,6 +16,6 @@ namespace ACL.Business.Application.Interfaces.Repositories
         /// <inheritdoc/>
         bool Delete(Branch entity);
         /// <inheritdoc/>
-        Branch? Delete(ulong id);
+        Branch? Delete(uint id);
     }
 }
