+++ ACL/src/ACL.Business/Application/Interfaces/Repositories/IPageRouteRepository.cs
@@ -11,7 +11,7 @@ namespace ACL.Business.Application.Interfaces.Repositories
         /// <inheritdoc/>
         List<PageRoute>? All();
         /// <inheritdoc/>
-        PageRoute? Find(ulong id);
+        PageRoute? Find(uint id);
         /// <inheritdoc/>
         PageRoute? Add(PageRoute pageRoute);
         /// <inheritdoc/>
@@ -19,8 +19,8 @@ namespace ACL.Business.Application.Interfaces.Repositories
         /// <inheritdoc/>
         PageRoute? Delete(PageRoute pageRoute);
         /// <inheritdoc/>
-        PageRoute? Delete(ulong id);
+        PageRoute? Delete(uint id);
         /// <inheritdoc/>
-        PageRoute[]? DeleteAllByPageId(ulong pageid);
+        PageRoute[]? DeleteAllByPageId(uint pageid);
     }
 }
