+++ ACL/src/ACL.Business/Application/Interfaces/Repositories/IStateRepository.cs
@@ -8,7 +8,7 @@ namespace ACL.Business.Application.Interfaces.Repositories
         /// <inheritdoc/>
         List<State>? All();
         /// <inheritdoc/>
-        State? Find(ulong id);
+        State? Find(uint id);
         /// <inheritdoc/>
         State? Add(State state);
         /// <inheritdoc/>
@@ -16,7 +16,7 @@ namespace ACL.Business.Application.Interfaces.Repositories
         /// <inheritdoc/>
         State? Delete(State state);
         /// <inheritdoc/>
-        State? Delete(ulong id);
+        State? Delete(uint id);
 
 
     }
