+++ ACL/src/ACL.Business/Contracts/Requests/AclPasswordResetRequest.cs
@@ -7,9 +7,9 @@ public partial class AclPasswordResetRequest
 {
     [DefaultValue("10")]
     [Required(ErrorMessage = "User ID is required")]
-    [Range(1, ulong.MaxValue, ErrorMessage = "User ID greater than 0.")]
+    [Range(1, uint.MaxValue, ErrorMessage = "User ID greater than 0.")]
     //[ExistsInDatabase<ApplicationDbContext,ICustomUnitOfWork>("User", "Id")]
-    public ulong UserId { get; set; }
+    public uint UserId { get; set; }
 
     [DefaultValue("12345678")]
     [Required(ErrorMessage = "Current password is required")]
