+++ ACL/src/ACL.Business/Domain/Entities/SubModule.cs
@@ -2,9 +2,9 @@
 
 public partial class SubModule
 {
-    public ulong Id { get; set; }
+    public uint Id { get; set; }
 
-    public ulong ModuleId { get; set; }
+    public uint ModuleId { get; set; }
 
     public string Name { get; set; } = null!;
