+++ ACL/src/ACL.Business/Domain/Entities/UserUsergroup.cs
@@ -2,13 +2,13 @@
 
 public partial class UserUsergroup
 {
-    public ulong Id { get; set; }
+    public uint Id { get; set; }
 
-    public ulong UsergroupId { get; set; }
+    public uint UsergroupId { get; set; }
 
-    public ulong UserId { get; set; }
+    public uint UserId { get; set; }
 
-    public ulong CompanyId { get; set; }
+    public uint CompanyId { get; set; }
 
     public DateTime? CreatedAt { get; set; }
